+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Ability to run tests without suites.
+
 ## [1.2.0] - 2026-05-21
 
 ### Added