+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ability to run tests without suites.
 
+### Changed
+
+- `testo.Options` now returns an empty struct to enable `var _ = testo.Options(...)` usage.
+
 ## [1.2.0] - 2026-05-21
 
 ### Added