+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- VS Code extension snippets for common Testo blocks.
 - Ability to run tests without suites.
 
 ### Changed